refactor : change directory
